refactor(sidebar): delete SidebarWrapper, inline into StepSidebarWrapper
