+++ web/src/sections/sidebar/StepSidebarWrapper.tsx
@@ -1,7 +1,9 @@
 import { ReactNode } from "react";
 import type { IconProps } from "@opal/types";
+import { SidebarLayouts } from "@opal/layouts";
 import { SidebarTab } from "@opal/components";
-import SidebarWrapper from "@/sections/sidebar/SidebarWrapper";
+import { renderAppLogo } from "@/lib/app/utils";
+import { useShowLogoWhenFolded } from "@/lib/sidebar/hooks";
 
 export interface StepSidebarProps {
   children: ReactNode;
@@ -16,8 +18,14 @@ export default function StepSidebar({
   buttonIcon,
   buttonHref,
 }: StepSidebarProps) {
+  const showLogoWhenFolded = useShowLogoWhenFolded();
+
   return (
-    <SidebarWrapper>
+    <SidebarLayouts.Root>
+      <SidebarLayouts.Header
+        logo={renderAppLogo}
+        showLogoWhenFolded={showLogoWhenFolded}
+      />
       <div className="px-2">
         <SidebarTab icon={buttonIcon} href={buttonHref}>
           {buttonName}
@@ -25,6 +33,6 @@ export default function StepSidebar({
       </div>
 
       <div className="h-full w-full px-4">{children}</div>
-    </SidebarWrapper>
+    </SidebarLayouts.Root>
   );
 }
